Refactor library management system
- Updated .gitattributes and .gitignore for better handling of line endings and added .devcontainer directory.
- Modified maven-wrapper.properties to include license information and updated Maven version.
- Enhanced README.md with author information.
- Refactored application.properties for better configuration management.
- Removed obsolete class files from output directories.
- Implemented DAO pattern with LibroDAO and UsuarioDAO for database interactions.
- Introduced models for Libro and Usuario with JPA annotations for ORM.
- Refactored existing classes to utilize new model structure and DAO pattern.
- Updated exception handling and repository methods for better error management.
- Cleaned up code formatting and ensured consistent use of imports.
