--- src/main/java/com/libreria/libreria/Usuario.java
@@ -1,61 +1,0 @@
-package com.libreria.libreria;
-
-public class Usuario {
-    private Long id;
-    private String nombre;
-    private String email;
-    private String password;
-
-    public Usuario() {
-    }
-
-    public Usuario(Long id, String nombre, String email, String password) {
-        this.id = id;
-        this.nombre = nombre;
-        setEmail(email);
-        setPassword(password);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        if (nombre == null || nombre.trim().isEmpty()) {
-            throw new IllegalArgumentException("El nombre no puede estar vacío.");
-        }
-        this.nombre = nombre;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        if (password == null || password.length() < 6) {
-            throw new IllegalArgumentException("La contraseña debe tener al menos 6 caracteres.");
-        }
-        this.password = password;
-    }
-
-    @Override
-    public String toString() {
-        return "User{id=" + id + ", nombre='" + nombre + "', email='" + email + "'}";
-    }
-}
