+++ src/main/java/com/libreria/libreria/models/Libro.java
@@ -1,0 +1,34 @@
+package com.libreria.libreria.models;
+
+import java.util.UUID;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+@Entity
+@Table(name = "libros")
+public class Libro {
+  @Id
+  @GeneratedValue(strategy = GenerationType.UUID)
+  private UUID isbn;
+  private String titulo;
+  private String autor;
+  private String genero;
+  private int anioPublicacion;
+  @Builder.Default
+  private boolean disponible = true;
+
+}
